requirements.txt
langchain~=0.0.316
streamlit~=1.27.2
python-dotenv~=1.0.0
pandas~=1.5.3
Pillow~=10.1.0
SQLAlchemy~=1.4.49
pandasai~=1.3.3